[🚀\ CHORE] API 목록 수정
